Do recursive function checks early during IR linking
