+++ source/slang/slang-ir-check-recursion.h
@@ -4,6 +4,10 @@ namespace Slang
 {
 struct IRModule;
 class DiagnosticSink;
+class TargetRequest;
 
 void checkForRecursiveTypes(IRModule* module, DiagnosticSink* sink);
+
+void checkForRecursiveFunctions(TargetRequest* target, IRModule* module, DiagnosticSink* sink);
+
 } // namespace Slang
