+++ source/slang/slang-ir-check-unsupported-inst.cpp
@@ -5,46 +5,6 @@
 
 namespace Slang
 {
-bool isCPUTarget(TargetRequest* targetReq);
-
-bool checkRecursionImpl(
-    HashSet<IRFunc*>& checkedFuncs,
-    HashSet<IRFunc*>& callStack,
-    IRFunc* func,
-    DiagnosticSink* sink)
-{
-    for (auto block : func->getBlocks())
-    {
-        for (auto inst : block->getChildren())
-        {
-            auto callInst = as<IRCall>(inst);
-            if (!callInst)
-                continue;
-            auto callee = as<IRFunc>(callInst->getCallee());
-            if (!callee)
-                continue;
-            if (!callStack.add(callee))
-            {
-                sink->diagnose(callInst, Diagnostics::unsupportedRecursion, callee);
-                return false;
-            }
-            if (checkedFuncs.add(callee))
-                checkRecursionImpl(checkedFuncs, callStack, callee, sink);
-            callStack.remove(callee);
-        }
-    }
-    return true;
-}
-
-void checkRecursion(HashSet<IRFunc*>& checkedFuncs, IRFunc* func, DiagnosticSink* sink)
-{
-    HashSet<IRFunc*> callStack;
-    if (checkedFuncs.add(func))
-    {
-        callStack.add(func);
-        checkRecursionImpl(checkedFuncs, callStack, func, sink);
-    }
-}
 
 void checkUnsupportedInst(TargetRequest* target, IRFunc* func, DiagnosticSink* sink)
 {
@@ -65,8 +25,6 @@ void checkUnsupportedInst(TargetRequest* target, IRFunc* func, DiagnosticSink* s
 
 void checkUnsupportedInst(TargetRequest* target, IRModule* module, DiagnosticSink* sink)
 {
-    HashSet<IRFunc*> checkedFuncsForRecursionDetection;
-
     for (auto globalInst : module->getGlobalInsts())
     {
         switch (globalInst->getOp())
@@ -84,8 +42,6 @@ void checkUnsupportedInst(TargetRequest* target, IRModule* module, DiagnosticSin
                 break;
             }
         case kIROp_Func:
-            if (!isCPUTarget(target))
-                checkRecursion(checkedFuncsForRecursionDetection, as<IRFunc>(globalInst), sink);
             checkUnsupportedInst(target, as<IRFunc>(globalInst), sink);
             break;
         case kIROp_Generic:
